pyproject.toml
[project]
name = "semantic-memory-agent"
version = "0.1.0"
description = "A personal semantic memory agent using ADK and Qdrant"
requires-python = ">=3.10"
dependencies = [
    "cohere>=5.20.0",
    "fastapi>=0.118.3",
    "google-genai>=1.0.0",
    "linkup-sdk>=0.9.0",
    "qdrant-client>=1.12.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.38.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
